+++ custom_components/cync_lights/switch.py
@@ -74,7 +74,11 @@ class CyncSwitchEntity(CoordinatorEntity[CyncCoordinator], SwitchEntity):
         pd = self._state.pycync_dev
         cc = getattr(pd, "_command_client", None)
         self.coordinator.note_command(self._switch_id, True)
-        if cc:
+        local = self.coordinator.local_command_target(self._switch_id)
+        srv = self.coordinator.local_server
+        if local and srv:
+            srv.set_power(local[0], local[1], True)
+        elif cc:
             await cc.set_power_state(pd, True)
         self._state.power = True
         self.async_write_ha_state()
@@ -83,7 +87,11 @@ class CyncSwitchEntity(CoordinatorEntity[CyncCoordinator], SwitchEntity):
         pd = self._state.pycync_dev
         cc = getattr(pd, "_command_client", None)
         self.coordinator.note_command(self._switch_id, False)
-        if cc:
+        local = self.coordinator.local_command_target(self._switch_id)
+        srv = self.coordinator.local_server
+        if local and srv:
+            srv.set_power(local[0], local[1], False)
+        elif cc:
             await cc.set_power_state(pd, False)
         self._state.power = False
         self.async_write_ha_state()